start refactor...
