+++ packages/apps/src/events/activity-response.ts
@@ -1,0 +1,19 @@
+import { Activity, ConversationReference, InvokeResponse } from '@microsoft/spark.api';
+
+import { IEvent } from '../types';
+
+/**
+ * the event emitted by a plugin
+ * before an invoke response is returned
+ */
+export interface IActivityResponseEvent extends IEvent, ConversationReference {
+  /**
+   * inbound request activity payload
+   */
+  activity: Activity;
+
+  /**
+   * the response
+   */
+  response: InvokeResponse;
+}
