+++ packages/apps/src/events/activity-sent.ts
@@ -1,0 +1,14 @@
+import { ActivityParams, ConversationReference } from '@microsoft/spark.api';
+
+import { IEvent } from '../types';
+
+/**
+ * the event emitted by a plugin
+ * when an activity is sent
+ */
+export interface IActivitySentEvent extends IEvent, ConversationReference {
+  /**
+   * the sent activity
+   */
+  activity: { id: string } & ActivityParams;
+}
