+++ packages/apps/src/events/index.ts
@@ -1,63 +1,26 @@
 import { ILogger } from '@microsoft/spark.common';
 
-import { IActivityErrorContext, IActivitySignInContext } from '../contexts';
-import {
-  IActivityBeforeSentEvent,
-  IActivityResponseEvent,
-  IActivityReceivedEvent,
-  IActivitySentEvent,
-} from '../types';
-import { ErrorEvent } from './error';
+import { IActivitySignInContext } from '../contexts';
 
-export interface IAppActivityErrorEvent extends IActivityErrorContext {
-  /**
-   * the unique name of the plugin that
-   * emitted the event
-   */
-  plugin: string;
-}
-
-export interface IAppActivityReceivedEvent extends IActivityReceivedEvent {
-  /**
-   * the unique name of the plugin that
-   * emitted the event
-   */
-  plugin: string;
-}
-
-export interface IAppActivityResponseEvent extends IActivityResponseEvent {
-  /**
-   * the unique name of the plugin that
-   * emitted the event
-   */
-  plugin: string;
-}
-
-export interface IAppActivitySentEvent extends IActivitySentEvent {
-  /**
-   * the unique name of the plugin that
-   * emitted the event
-   */
-  plugin: string;
-}
-
-export interface IAppActivityBeforeSentEvent extends IActivityBeforeSentEvent {
-  /**
-   * the unique name of the plugin that
-   * emitted the event
-   */
-  plugin: string;
-}
+import { IActivityBeforeSentEvent } from './activity-before-sent';
+import { IActivityReceivedEvent } from './activity-received';
+import { IActivityResponseEvent } from './activity-response';
+import { IActivitySentEvent } from './activity-sent';
+import { IErrorEvent, IActivityErrorEvent } from './error';
 
 export interface IEvents {
   start: ILogger;
   signin: IActivitySignInContext;
-  error: ErrorEvent;
-  'activity.error': IAppActivityErrorEvent;
-  'activity.received': IAppActivityReceivedEvent;
-  'activity.response': IAppActivityResponseEvent;
-  'activity.sent': IAppActivitySentEvent;
-  'activity.before.sent': IAppActivityBeforeSentEvent;
+  error: IErrorEvent;
+  'activity.error': IActivityErrorEvent;
+  'activity.received': IActivityReceivedEvent;
+  'activity.response': IActivityResponseEvent;
+  'activity.sent': IActivitySentEvent;
+  'activity.before.sent': IActivityBeforeSentEvent;
 }
 
-export { ErrorEvent };
+export * from './activity-before-sent';
+export * from './activity-received';
+export * from './activity-response';
+export * from './activity-sent';
+export * from './error';
