--- packages/apps/src/events/signin.ts
@@ -1,5 +1,0 @@
-import { IActivitySignInContext } from '../contexts';
-
-export function signin({ activity, log }: IActivitySignInContext) {
-  log.debug(`${activity.from.name} has signed in...`);
-}
