+++ packages/apps/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './plugin';
 export * from './streamer';
 export * from './event-handler';
+export * from './event';
 export * from './route-handler';
 export * from './prefixed';
 export * from './suffixed';
